Define jump power in terms of jump height
Replace jump_force -> jump_height so we can tune with a number that's
relevant to gameplay: use the size of a tile or distance between two
platforms.

Could precompute jump_height, but you wouldn't be able to tune it and
jump_gravity_acceleration while playing. Might be better to use setget,
but I can't manage to keep it in sync with jump_gravity_acceleration and
it seems overcomplicated for this project?

Credit to this tutorial by AuraTummyache:
https://www.reddit.com/r/godot/comments/sioxoc/ive_seen_a_lot_of_people_doing_jumps_with/

This link goes further into the projectile math and also has a solution
to tune gravity in terms of time and jump_height:
https://medium.com/@brazmogu/physics-for-game-dev-a-platformer-physics-cheatsheet-f34b09064558

I solved the equation to find jump_height based on the previous values
to keep the tuning the same:

    jump_force = -sqrt(2 * jump_gravity_acceleration * jump_height)
    (-jump_force)^2 / 2 / jump_gravity_acceleration = jump_height
    jump_height = 1300*1300/2/4000
    jump_height = 211.25

Test
* Placed a Marker2D on the ground and a child Sprite2D 211 units above
  it. Hold jump and I reach the sprite.