Rename gravity_max -> max_drop_speed_for_gravity
This is not a gravity value. It's a threshold for whether we should make
you fall faster.

Rename to make it clearer that it's related to velocity and not gravity.
Our current value of 1000 makes it unlikely to hit unless you have a
large fall.

Use a range to make it clear that it must be positive.